agent: cross-platform runner + Windows (WinMM) backend; Windows CI
Refactors the daemon into a shared, platform-agnostic runner driven by a
small MIDI-source interface, then adds the Windows backend - the user's
actual ask (a Windows agent; the tray GUI layers on this next).

- agent/runner.{c,h}: the whole loop (connect, fetch bindings, match,
  coalesce knobs, POST /api/knob & /api/trigger, refresh, reconnect,
  --dry-run, verbose) lives here ONCE, pulling MIDI through tm_midi_src
  {read,wait}. One dispatch implementation for every platform.
- agent/net/net.{c,h}: moved from linux/ and made cross-platform
  (POSIX sockets / Winsock via #ifdef _WIN32; intptr_t socket handle).
- agent/windows/midi_win.{c,h}: WinMM input - enumerate devices by
  friendly name, decode short messages in the callback into a ring the
  runner drains; event-signalled wait.
- agent/windows/main_win.c: Windows console entry (same --list/--dry-run/
  name-substring UX as Linux); Ctrl-C -> tm_runner_stop.
- agent/midi_port.h: shared port struct; linux/main.c is now a thin ALSA
  source over the runner.
- CMake: shared runner+net compiled into each platform target; Windows
  target links winmm + ws2_32; MSVC secure-CRT defines.
- CI: agent-core now builds+tests on BOTH ubuntu (gcc+ALSA) and
  windows-latest (MSVC+WinMM) - real cross-platform validation without a
  local compiler.

Portable core untouched. HTTP transport unchanged.

+++ agent/net/net.h
@@ -1,11 +1,13 @@
-/* net.h — minimal HTTP/1.1 keep-alive client over POSIX sockets (Linux/macOS).
- * Just enough for the agent: one request at a time, Content-Length responses.
- * No TLS (the bridge is plain HTTP on the LAN); portable to Winsock later. */
+/* net.h — minimal HTTP/1.1 keep-alive client. Cross-platform: POSIX sockets on
+ * Linux/macOS, Winsock on Windows. Just enough for the agent: one request at a
+ * time, Content-Length responses, no TLS (the bridge is plain HTTP on the LAN). */
 #ifndef TM_NET_H
 #define TM_NET_H
 
+#include <stdint.h>
+
 typedef struct {
-    int fd;
+    intptr_t fd;      /* socket handle (SOCKET on Windows fits in intptr_t) */
     char host[128];
     int port;
 } tm_net;
